Add files via upload
The sorted results (variables/predictors/features selected) from running a Backward Elimination Stepwise Regression on all 58,500 datasets organized in the form of two columns with headers and 58,501 rows. The first column has the name of the csv file with the dataset in it and the second column has the variables selected by BE Stepwise for the corresponding dataset whose name appears in the 1st column.