Add files via upload
If anyone looking at this Repository ever wants to truly reproduce my findings, you are going to have to create all of the datasets yourself (or you can email me at spencermstarr@hotmail.com and I'll share access of a zipped version of the folder with all 260k datasets in it I have in my Google Drive). And if you do, for whatever reason, about half a dozen of the 260k csv file formatted datasets it created might be blank which breaks my R scripts, so you will have to do what I did that is described in detail here as a back and forth with Chat GPT 4.0 on how to alter the VBA code in custom macro in the 'Data Generator' Macro Enabled Excel file I have uploaded in this Repo.