Add files via upload
These 4 scripts were actually the cleverest of all the ideas I had to help me save time whenever I had to re-run any of the benchmarks because I had spotted an error in the results, or realized I had forgotten to add the code to calculate another classification model performance metric for this benchmark which I had just recently added into the scripts for all the others and run them again. 

The reason they are such a huge time saver is because they part of this project with the longest run times is importing and transforming the 260k datasets in sub sets of 5,000, 10,000, or 15,000 at most at a time (and doing so for a subset of 10,000 takes more than twice as long as doing so for a subset of only 5,000), and then saving that WorkSpace, which takes a minute or two so it can be loaded later which only takes a few seconds. Using these 4 scripts, you can take 3 consecutive saved WorkSpaces, each of which contain 5k datasets ready to have BM 1, 2, or 3 run on it, load them into RStudio one at a time, starting off with Renaming Environment Objects script A.R in which you can save every object in it to an object with the same name to what it is saved as to that same name followed by an A or 1, then remove the original name, then repeat for the 2nd WorkSpace with names followed by B or 2 using the Renaming Environment Objects script B script.R, and the 3rd with the same names followed by C or 3.

Then, after all of that, using Merging datasets.R, you can easily rename them all back to their original names so they can be run using the rest of the scripts. This might sound complicated, but it only takes a couple of minutes and can help you save a tremendous amount of time by combining datasets before running for example, LASSO using the lars() function which runs extremely quickly.